# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Elrond Portal API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://elrond-portal.test"
        }
    ],
    "paths": {
...
        "/api/v1/council": {
            "get": {
                "tags": [
                    "Council"
                ],
                "summary": "Retorna informações dos conselhos",
                "operationId": "7a4e39c5fa3ea47dfe06e31fafa73149",
                "responses": {
                    "200": {
                        "description": "Informações dos conselhos retornadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "councils": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Conselho Exemplo"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descrição do Conselho Exemplo"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Council"
                ],
                "summary": "Cria um novo conselho",
                "operationId": "3be6e5f0f5b0e9698b9f47e84c67bde0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Conselho Exemplo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição do Conselho Exemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conselho criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "council": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Conselho Exemplo",
                                                "description": "Descrição do Conselho Exemplo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
        "/api/v1/expedition-status": {
            "get": {
                "tags": [
                    "Expedition Status"
                ],
                "summary": "Retorna informações dos status de expedições",
                "operationId": "f0c7720d9142c87af3fc16fd8f0bd17e",
                "responses": {
                    "200": {
                        "description": "Informações dos status de expedições retornadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "expedition_statuses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Status Exemplo"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descrição do Status Exemplo"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "status-exemplo"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expedition Status"
                ],
                "summary": "Cria um novo status de expedição",
                "operationId": "a8ead08be54321caef4d71c6c6190b75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Status Exemplo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição do Status Exemplo"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "status-exemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status de expedição criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "expedition_status": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Status Exemplo",
                                                "description": "Descrição do Status Exemplo",
                                                "slug": "status-exemplo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}